pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Educational project: gendiff"
authors = ["Mikhail Davydov <mdavydov83@gmail.com>"]

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pip = "^22.3.1"
pytest = "^7.2.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"